Move append methods, reading a file works :tada:
