+++ astrodata/core.py
@@ -1,7 +1,8 @@
 import inspect
+import logging
 import os
 import re
-from collections import namedtuple
+from collections import namedtuple, OrderedDict
 from contextlib import suppress
 from copy import deepcopy
 from functools import partial, wraps
@@ -9,9 +10,12 @@ from functools import partial, wraps
 import numpy as np
 
 from astropy.io import fits
+from astropy.nddata import NDData
 from astropy.table import Table
 
-from .nddata import NDAstroData as NDDataObject, ADVarianceUncertainty
+from .fits import DEFAULT_EXTENSION, _process_table, read_fits, write_fits
+from .nddata import ADVarianceUncertainty
+from .nddata import NDAstroData as NDDataObject
 from .utils import deprecated, normalize_indices
 
 NO_DEFAULT = object()
@@ -194,9 +198,12 @@ class AstroData:
         'ut_date': 'DATE-OBS'
     }
 
-    def __init__(self, nddatas, other=None, phu=None, indices=None):
-        if isinstance(nddatas, (list, tuple)):
-            nddatas = dict(zip(range(len(nddatas)), nddatas))
+    def __init__(self, nddatas=None, other=None, phu=None, indices=None):
+        if nddatas is None:
+            nddatas = []
+        if not isinstance(nddatas, (list, tuple)):
+            nddatas = list(nddatas)
+
         self._nddata = nddatas
         self._other = other
         self._phu = phu or fits.Header()
@@ -205,6 +212,7 @@ class AstroData:
         self._exposed = set()
         self._fixed_settable = {'data', 'uncertainty', 'mask', 'variance',
                                 'wcs', 'path', 'filename'}
+        self._logger = logging.getLogger(__name__)
         self._orig_filename = None
         self._path = None
         self._processing_tags = False
@@ -450,6 +458,10 @@ class AstroData:
         else:
             ndd = self._nddata
         return ndd[0] if self.is_single else ndd
+
+    def table(self):
+        # FIXME: do we need this in addition to .tables ?
+        return self._tables.copy()
 
     @property
     def tables(self):
@@ -740,49 +752,6 @@ class AstroData:
             return len(self._indices)
         else:
             return len(self._nddata)
-
-    def append(self, ext, name=None, **kwargs):
-        """
-        Adds a new top-level extension. Objects appended to a single
-        slice will actually be made hierarchically dependent of the science
-        object represented by that slice. If appended to the provider as
-        a whole, the new member will be independent (eg. global table, new
-        science object).
-
-        Args
-        -----
-        ext : array, `NDData`, `Table`, other
-            The contents for the new extension. The exact accepted types depend
-            on the class implementing this interface. Implementations specific
-            to certain data formats may accept specialized types (eg. a FITS
-            provider will accept an `ImageHDU` and extract the array out of it)
-
-        name : str, optional
-            A name that may be used to access the new object, as an attribute
-            of the provider. The name is typically ignored for top-level
-            (global) objects, and required for the others. If the name cannot
-            be derived from the metadata associated to `extension`, you will
-            have to provider one.
-
-            It can consist in a combination of numbers and letters, with the
-            restriction that the letters have to be all capital, and the first
-            character cannot be a number ("[A-Z][A-Z0-9]*").
-
-        Returns
-        --------
-        The same object, or a new one, if it was necessary to convert it to
-        a more suitable format for internal use.
-
-        Raises
-        -------
-        TypeError
-            If adding the object in an invalid situation (eg. `name` is `None`
-            when adding to a single slice)
-        ValueError
-            Raised if the extension is of a proper type, but its value is
-            illegal somehow.
-
-        """
 
     @property
     def exposed(self):
@@ -1092,12 +1061,278 @@ class AstroData:
         copy._oper(copy._rdiv, oper)
         return copy
 
+    def _reset_ver(self, nd):
+        try:
+            ver = max(_nd.meta['ver'] for _nd in self._nddata) + 1
+        except ValueError:
+            # This seems to be the first extension!
+            ver = 1
+
+        nd.meta['header']['EXTVER'] = ver
+        nd.meta['ver'] = ver
+
+        try:
+            oheaders = nd.meta['other_header']
+            for extname, ext in nd.meta['other'].items():
+                try:
+                    oheaders[extname]['EXTVER'] = ver
+                except KeyError:
+                    pass
+                try:
+                    # The object may keep the header on its own structure
+                    ext.meta['header']['EXTVER'] = ver
+                except AttributeError:
+                    pass
+        except KeyError:
+            pass
+
+        return ver
+
+    def _process_pixel_plane(self, pixim, name=None, top_level=False,
+                             reset_ver=True, custom_header=None):
+        if not isinstance(pixim, NDDataObject):
+            # Assume that we get an ImageHDU or something that can be
+            # turned into one
+            if isinstance(pixim, fits.ImageHDU):
+                nd = NDDataObject(pixim.data, meta={'header': pixim.header})
+            elif custom_header is not None:
+                nd = NDDataObject(pixim, meta={'header': custom_header})
+            else:
+                nd = NDDataObject(pixim, meta={'header': {}})
+        else:
+            nd = pixim
+            if custom_header is not None:
+                nd.meta['header'] = custom_header
+
+        header = nd.meta['header']
+        currname = header.get('EXTNAME')
+        ver = header.get('EXTVER', -1)
+
+        # TODO: Review the logic. This one seems bogus
+        if name and (currname is None):
+            header['EXTNAME'] = name if name is not None else DEFAULT_EXTENSION
+
+        if top_level:
+            if 'other' not in nd.meta:
+                nd.meta['other'] = OrderedDict()
+                nd.meta['other_header'] = {}
+
+            if reset_ver or ver == -1:
+                self._reset_ver(nd)
+            else:
+                nd.meta['ver'] = ver
+
+        return nd
+
+    def _add_to_other(self, add_to, name, data, header=None):
+        meta = add_to.meta
+        meta['other'][name] = data
+        if header:
+            header['EXTVER'] = meta.get('ver', -1)
+            meta['other_header'][name] = header
+
+    def _append_array(self, data, name=None, header=None, add_to=None):
+        if add_to is None:
+            # Top level extension
+
+            # Special cases for Gemini
+            if name is None:
+                name = DEFAULT_EXTENSION
+
+            if name in {'DQ', 'VAR'}:
+                raise ValueError("'{}' need to be associated to a '{}' one"
+                                 .format(name, DEFAULT_EXTENSION))
+            else:
+                # FIXME: the logic here is broken since name is
+                # always set to somehing above with DEFAULT_EXTENSION
+                if name is not None:
+                    hname = name
+                elif header is not None:
+                    hname = header.get('EXTNAME', DEFAULT_EXTENSION)
+                else:
+                    hname = DEFAULT_EXTENSION
+
+                hdu = fits.ImageHDU(data, header=header)
+                hdu.header['EXTNAME'] = hname
+                ret = self._append_imagehdu(hdu, name=hname, header=None,
+                                            add_to=None)
+        else:
+            # Attaching to another extension
+            if header is not None and name in {'DQ', 'VAR'}:
+                self._logger.warning(
+                    "The header is ignored for '{}' extensions".format(name))
+            if name is None:
+                raise ValueError("Can't append pixel planes to other "
+                                 "objects without a name")
+            elif name is DEFAULT_EXTENSION:
+                raise ValueError("Can't attach '{}' arrays to other objects"
+                                 .format(DEFAULT_EXTENSION))
+            elif name == 'DQ':
+                add_to.mask = data
+                ret = data
+            elif name == 'VAR':
+                std_un = ADVarianceUncertainty(data)
+                std_un.parent_nddata = add_to
+                add_to.uncertainty = std_un
+                ret = std_un
+            else:
+                self._add_to_other(add_to, name, data, header=header)
+                ret = data
+
+        return ret
+
+    def _append_imagehdu(self, hdu, name, header, add_to, reset_ver=True):
+        if name in {'DQ', 'VAR'} or add_to is not None:
+            return self._append_array(hdu.data, name=name, add_to=add_to)
+        else:
+            nd = self._process_pixel_plane(hdu, name=name, top_level=True,
+                                           reset_ver=reset_ver,
+                                           custom_header=header)
+            return self._append_nddata(nd, name, add_to=None)
+
+    def _append_raw_nddata(self, raw_nddata, name, header, add_to,
+                           reset_ver=True):
+        # We want to make sure that the instance we add is whatever we specify
+        # as `NDDataObject`, instead of the random one that the user may pass
+        top_level = add_to is None
+        if not isinstance(raw_nddata, NDDataObject):
+            raw_nddata = NDDataObject(raw_nddata)
+        processed_nddata = self._process_pixel_plane(raw_nddata,
+                                                     top_level=top_level,
+                                                     custom_header=header,
+                                                     reset_ver=reset_ver)
+        return self._append_nddata(processed_nddata, name=name, add_to=add_to)
+
+    def _append_nddata(self, new_nddata, name, add_to, reset_ver=True):
+        # NOTE: This method is only used by others that have constructed NDData
+        # according to our internal format. We don't accept new headers at this
+        # point, and that's why it's missing from the signature.  'name' is
+        # ignored. It's there just to comply with the _append_XXX signature.
+        if add_to is not None:
+            raise TypeError("You can only append NDData derived instances "
+                            "at the top level")
+
+        hd = new_nddata.meta['header']
+        hname = hd.get('EXTNAME', DEFAULT_EXTENSION)
+        if hname == DEFAULT_EXTENSION:
+            if reset_ver:
+                self._reset_ver(new_nddata)
+            self._nddata.append(new_nddata)
+        else:
+            raise ValueError("Arbitrary image extensions can only be added "
+                             "in association to a '{}'"
+                             .format(DEFAULT_EXTENSION))
+
+        return new_nddata
+
+    def _append_table(self, new_table, name, header, add_to, reset_ver=True):
+        tb = _process_table(new_table, name, header)
+        hname = tb.meta['header'].get('EXTNAME') if name is None else name
+        # if hname is None:
+        #     raise ValueError("Can't attach a table without a name!")
+        if add_to is None:
+            if hname is None:
+                table_num = 1
+                while 'TABLE{}'.format(table_num) in self._tables:
+                    table_num += 1
+                hname = 'TABLE{}'.format(table_num)
+            # Don't use setattr, which is overloaded and may case problems
+            self.__dict__[hname] = tb
+            self._tables[hname] = tb
+            self._exposed.add(hname)
+        else:
+            if hname is None:
+                table_num = 1
+                while getattr(add_to, 'TABLE{}'.format(table_num), None):
+                    table_num += 1
+                hname = 'TABLE{}'.format(table_num)
+            setattr(add_to, hname, tb)
+            self._add_to_other(add_to, hname, tb, tb.meta['header'])
+            add_to.meta['other'][hname] = tb
+        return tb
+
+    def _append_astrodata(self, ad, name, header, add_to, reset_ver=True):
+        if not ad.is_single:
+            raise ValueError("Cannot append AstroData instances that are "
+                             "not single slices")
+        elif add_to is not None:
+            raise ValueError("Cannot append an AstroData slice to "
+                             "another slice")
+
+        new_nddata = deepcopy(ad.nddata)
+        if header is not None:
+            new_nddata.meta['header'] = deepcopy(header)
+
+        return self._append_nddata(new_nddata, name=None, add_to=None,
+                                   reset_ver=True)
+
+    def append(self, ext, name=None, header=None, reset_ver=True, add_to=None):
+        """
+        Adds a new top-level extension. Objects appended to a single
+        slice will actually be made hierarchically dependent of the science
+        object represented by that slice. If appended to the provider as
+        a whole, the new member will be independent (eg. global table, new
+        science object).
+
+        Args
+        -----
+        ext : array, `NDData`, `Table`, other
+            The contents for the new extension. The exact accepted types depend
+            on the class implementing this interface. Implementations specific
+            to certain data formats may accept specialized types (eg. a FITS
+            provider will accept an `ImageHDU` and extract the array out of it)
+
+        name : str, optional
+            A name that may be used to access the new object, as an attribute
+            of the provider. The name is typically ignored for top-level
+            (global) objects, and required for the others. If the name cannot
+            be derived from the metadata associated to `extension`, you will
+            have to provider one.
+
+            It can consist in a combination of numbers and letters, with the
+            restriction that the letters have to be all capital, and the first
+            character cannot be a number ("[A-Z][A-Z0-9]*").
+
+        Returns
+        --------
+        The same object, or a new one, if it was necessary to convert it to
+        a more suitable format for internal use.
+
+        Raises
+        -------
+        TypeError
+            If adding the object in an invalid situation (eg. `name` is `None`
+            when adding to a single slice)
+        ValueError
+            Raised if the extension is of a proper type, but its value is
+            illegal somehow.
+
+        """
+        # NOTE: Most probably, if we want to copy the input argument, we
+        #       should do it here...
+        if isinstance(ext, fits.PrimaryHDU):
+            raise ValueError("Only one Primary HDU allowed. "
+                             "Use set_phu if you really need to set one")
+
+        dispatcher = (
+            (NDData, self._append_raw_nddata),
+            ((Table, fits.TableHDU, fits.BinTableHDU), self._append_table),
+            (fits.ImageHDU, self._append_imagehdu),
+            (AstroData, self._append_astrodata),
+        )
+
+        for bases, method in dispatcher:
+            if isinstance(ext, bases):
+                return method(ext, name=name, header=header, add_to=add_to,
+                              reset_ver=reset_ver)
+        else:
+            # Assume that this is an array for a pixel plane
+            return self._append_array(ext, name=name, header=header,
+                                      add_to=add_to)
+
     @classmethod
     def read(cls, source):
-        """
-        Read from a file, file object, HDUList, etc.
-        """
-        from .fits import read_fits
+        """Read from a file, file object, HDUList, etc."""
         return read_fits(cls, source)
 
     load = read  # for backward compatibility
@@ -1108,7 +1343,6 @@ class AstroData:
                 raise ValueError("A filename needs to be specified")
             filename = self.path
 
-        from .fits import write_fits
         write_fits(self, filename, overwrite=overwrite)
 
     def extver_map(self):
@@ -1337,36 +1571,19 @@ class AstroData:
 
     @astro_data_descriptor
     def instrument(self):
-        """
-        Returns the name of the instrument making the observation
-
-        Returns
-        -------
-        str
-            instrument name
-        """
+        """Returns the name of the instrument making the observation."""
         return self.phu.get(self._keyword_for('instrument'))
 
     @astro_data_descriptor
     def object(self):
-        """
-        Returns the name of the object being observed
-
-        Returns
-        -------
-        str
-            object name
-        """
+        """Returns the name of the object being observed."""
         return self.phu.get(self._keyword_for('object'))
 
     @astro_data_descriptor
     def telescope(self):
-        """
-        Returns the name of the telescope
-
-        Returns
-        -------
-        str
-            name of the telescope
-        """
+        """Returns the name of the telescope."""
         return self.phu.get(self._keyword_for('telescope'))
+
+
+class AstroDataFits(AstroData):
+    """Keep this for now as other classes inherit from it."""
